fix(notifications): stabilize gateway build
