+++ packages/notifications-gateway/src/redis/schema.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import type { NotificationEnvelope } from '@agyn/shared';
 import { RoomSchema } from '../rooms';
 
-export const NotificationEnvelopeSchema = z
+export const NotificationEnvelopeSchema: z.ZodType<NotificationEnvelope> = z
   .object({
     id: z.string().min(1),
     ts: z.string().datetime(),
@@ -11,4 +11,4 @@ export const NotificationEnvelopeSchema = z
     event: z.string().min(1),
     payload: z.unknown(),
   })
-  .strict() satisfies z.ZodType<NotificationEnvelope>;
+  .strict();
