+++ packages/platform-server/__tests__/memory.runtime.integration.test.ts
@@ -5,7 +5,7 @@ import { LiveGraphRuntime } from '../src/graph-core/liveGraph.manager';
 import type { LLMContext } from '../src/llm/types';
 import { Signal } from '../src/signal';
 import { TemplateRegistry } from '../src/graph-core/templateRegistry';
-import type { GraphDefinition } from '../src/shared/types/graph.types';
+import type { GraphDefinition } from '@agyn/shared';
 import Node from '../src/nodes/base/Node';
 import { GraphRepository } from '../src/graph/graph.repository';
 import {
