+++ packages/platform-server/src/graph/services/graphVariables.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { PrismaService } from '../../core/services/prisma.service';
 import { GraphRepository } from '../graph.repository';
-import type { PersistedGraph } from '../../shared/types/graph.types';
+import type { PersistedGraph } from '@agyn/shared';
 
 export type VarItem = { key: string; graph: string | null; local: string | null };
