+++ packages/platform-server/src/shared/shared.module.ts
@@ -2,6 +2,3 @@ import { Module } from '@nestjs/common';
 
 @Module({})
 export class SharedModule {}
-
-export * from './types/graph.types';
-export * from './types/common.types';
