+++ packages/shared/src/index.ts
@@ -1,23 +1,4 @@
-// Re-export selected shared types from server graph layer to avoid duplication
-export type {
-  TemplateNodeSchema,
-  PersistedGraph,
-  PersistedGraphNode,
-  PersistedGraphEdge,
-  PersistedGraphUpsertRequest,
-  PersistedGraphUpsertResponse,
-} from '../../platform-server/src/shared/types/graph.types';
-
-export type {
-  SecretRef,
-  VariableRef,
-  Reference,
-  ReferenceSource,
-  ReferenceValue,
-  ResolutionEvent,
-  ResolutionReport,
-  ResolutionEventSource,
-  ResolutionErrorCode,
-} from './references';
-
+export * from './graph';
+export * from './common';
+export * from './references';
 export * from './notifications';
